ofz: Integer-overflow
convert to double earlier

Reviewed-on: https://gerrit.libreoffice.org/c/core/+/112944
Tested-by: Jenkins

+++ svx/source/customshapes/EnhancedCustomShape2d.cxx
@@ -832,8 +832,8 @@ double EnhancedCustomShape2d::GetEnumFunc( const ExpressionFunct eFunc ) const
         case ExpressionFunct::EnumPi :         fRet = F_PI; break;
         case ExpressionFunct::EnumLeft :       fRet = static_cast<double>(nCoordLeft); break;
         case ExpressionFunct::EnumTop :        fRet = static_cast<double>(nCoordTop); break;
-        case ExpressionFunct::EnumRight :      fRet = static_cast<double>(nCoordLeft + nCoordWidth) * fXRatio; break;
-        case ExpressionFunct::EnumBottom :     fRet = static_cast<double>(nCoordTop + nCoordHeight) * fYRatio; break;
+        case ExpressionFunct::EnumRight :      fRet = (static_cast<double>(nCoordLeft) + static_cast<double>(nCoordWidth)) * fXRatio; break;
+        case ExpressionFunct::EnumBottom :     fRet = (static_cast<double>(nCoordTop) + static_cast<double>(nCoordHeight)) * fYRatio; break;
         case ExpressionFunct::EnumXStretch :   fRet = nXRef; break;
         case ExpressionFunct::EnumYStretch :   fRet = nYRef; break;
         case ExpressionFunct::EnumHasStroke :  fRet = bStroked ? 1.0 : 0.0; break;
